+++ src/app/components/reporting/realtime/realtimeDashboardManagement/versions/dashboardVersionsDirective.js
@@ -1,0 +1,13 @@
+'use strict';
+
+angular.module('liveopsConfigPanel')
+  .directive('dashboardVersions', [function() {
+    return {
+      scope: {
+        dashboard: '=',
+        onEditClick: '&'
+      },
+      templateUrl: 'app/components/reporting/realtime/versions/dashboardVersions.html',
+      controller: 'DashboardVersionsController'
+    };
+  }]);