+++ src/app/components/reporting/realtime/realtimeDashboardsManagementController.js
@@ -1,8 +1,0 @@
-'use strict';
-
-angular.module('liveopsConfigPanel').controller('RealtimeDashboardsManagementController', ['$scope', 'realtimeDashboardsManagementTableConfig', 'dashboards',
-  function($scope, realtimeDashboardsManagementTableConfig, dashboards) {
-    $scope.realtimeDashboardsManagementTableConfig = realtimeDashboardsManagementTableConfig;
-    $scope.dashboards = dashboards;
-  }
-]);